+++ crates/authly-client/src/error.rs
@@ -31,8 +31,8 @@ pub enum Error {
     EnvironmentNotInferrable,
 
     /// Invalid Common Name in certificate signing request.
-    #[error("invalid X509 common name")]
-    InvalidCommonName,
+    #[error("invalid X509 alt names")]
+    InvalidAltNames,
 
     /// A party was not authenticated or an operation was forbidden.
     #[error("unauthorized: {0}")]