+++ crates/authly-client/src/identity.rs
@@ -1,7 +1,8 @@
 //! Client identity, in the TLS sense.
 
-use std::borrow::Cow;
+use std::{borrow::Cow, str::FromStr};
 
+use authly_common::id::Eid;
 use pem::{EncodeConfig, Pem};
 
 use crate::Error;
@@ -77,4 +78,41 @@ impl Identity {
         identity_pem.extend(&self.key_pem);
         Ok(Cow::Owned(identity_pem))
     }
+}
+
+#[derive(Clone)]
+pub(crate) struct IdentityData {
+    pub entity_id: Eid,
+}
+
+pub(crate) fn parse_identity_data(cert: &[u8]) -> Result<IdentityData, Error> {
+    let pem = pem::parse(cert).map_err(|_| Error::AuthlyCA("invalid authly certificate"))?;
+
+    let (_, x509_cert) = x509_parser::parse_x509_certificate(pem.contents())
+        .map_err(|_| Error::AuthlyCA("invalid authly certificate"))?;
+
+    let mut entity_id: Option<Eid> = None;
+
+    for subject_attr in x509_cert.subject().iter_attributes() {
+        if let Some(oid_iter) = subject_attr.attr_type().iter() {
+            if oid_iter.eq(authly_common::certificate::oid::ENTITY_UNIQUE_IDENTIFIER
+                .iter()
+                .copied())
+            {
+                let value = subject_attr
+                    .attr_value()
+                    .as_str()
+                    .map_err(|_| Error::Identity("Entity Id value encoding"))?;
+                entity_id = Some(
+                    Eid::from_str(value)
+                        .map_err(|_| Error::Identity("Entity Id value encoding"))?,
+                );
+            }
+        }
+    }
+
+    let entity_id = entity_id.ok_or_else(|| Error::Identity("Entity Id is missing"))?;
+
+    // Assume that EC is always used
+    Ok(IdentityData { entity_id })
 }